migration/requirements.txt
psycopg[binary]>=3.1
mysql-connector-python>=8.0
bcrypt>=4.0
python-dotenv>=1.0
